Refactor asset.entity.ts to change the type of the 'inventory' column to 'int8'

+++ src/entities/asset.entity.ts
@@ -148,7 +148,7 @@ export class Asset {
     petIndex: number
 
     @Column({
-        type: 'smallint',
+        type: 'int8',
     })
     inventory: number
